# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Real-time Chat & Video Call API with Go, Gin, WebSocket, Redis Pub/Sub.",
        "version": "1.0",
        "title": "GoTalk API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@gotalk.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "api.localhost",
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/conversations": {
            "get": {
                "summary": "Get Conversations",
                "description": "Get list of conversations for current user.",
                "tags": [
                    "Conversations"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of conversations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ConversationResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Conversation",
                "description": "Create a new private or group conversation.",
                "tags": [
                    "Conversations"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Create conversation request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Conversation created",
                        "schema": {
                            "$ref": "#/definitions/ConversationResponse"
                        }
                    }
                }
            }
        },
        "/conversations/direct": {
            "post": {
                "summary": "Get or Create Direct Conversation",
                "description": "Find existing private chat with user, or create new one. Returns conversation + messages.",
                "tags": [
                    "Conversations"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Partner ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DirectConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Direct conversation info",
                        "schema": {
                            "$ref": "#/definitions/DirectConversationResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{id}/messages": {
            "get": {
                "summary": "Get Messages",
                "description": "Get messages in a conversation with pagination.",
                "tags": [
                    "Messages"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Conversation ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of messages to return",
                        "type": "integer",
                        "default": 50
                    },
                    {
                        "name": "before",
                        "in": "query",
                        "description": "Cursor (Message ID) to fetch messages before",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Message"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Send Message",
                "description": "Send a new message to a conversation.",
                "tags": [
                    "Messages"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Conversation ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Message content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Message sent",
                        "schema": {
                            "$ref": "#/definitions/Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CreateConversationRequest": {
            "type": "object",
            "required": [
                "type",
                "member_ids"
            ],
            "properties": {
                "type": {
                    "type": "string",
                    "enum": [
                        "private",
                        "group"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "member_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "DirectConversationRequest": {
            "type": "object",
            "required": [
                "partner_id"
            ],
            "properties": {
                "partner_id": {
                    "type": "string"
                }
            }
        },
        "DirectConversationResponse": {
            "type": "object",
            "properties": {
                "conversation": {
                    "$ref": "#/definitions/ConversationResponse"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Message"
                    }
                },
                "is_new": {
                    "type": "boolean"
                }
            }
        },
        "SendMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "text",
                        "image",
                        "file"
                    ],
                    "default": "text"
                },
                "file_url": {
                    "type": "string"
                },
                "reply_to_id": {
                    "type": "string"
                }
            }
        },
...
        "ConversationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unread_count": {
                    "type": "integer"
                },
                "last_message": {
                    "$ref": "#/definitions/Message"
                }
            }
        },
        "Message": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}